Refactored world package and removed world test
