+++ src/main/java/com/gorman/voxel_engine/world/primitives/Voxel.java
@@ -1,4 +1,4 @@
-package com.gorman.voxel_engine.world;
+package com.gorman.voxel_engine.world.primitives;
 
 import java.awt.Color;
