+++ src/main/java/com/gorman/voxel_engine/world/terrain/FlatTerrain.java
@@ -2,9 +2,8 @@ package com.gorman.voxel_engine.world.terrain;
 
 import java.awt.Color;
 
-import com.gorman.voxel_engine.world.Chunk;
-import com.gorman.voxel_engine.world.Vector;
-import com.gorman.voxel_engine.world.Voxel;
+import com.gorman.voxel_engine.world.primitives.Vector;
+import com.gorman.voxel_engine.world.primitives.Voxel;
 
 public class FlatTerrain extends Terrain {
