+++ src/test/java/com/gorman/voxel_engine/world/terrain/TestChunk.java
@@ -1,4 +1,4 @@
-package com.gorman.voxel_engine.world;
+package com.gorman.voxel_engine.world.terrain;
 
 import static org.junit.Assert.fail;
 import static org.junit.Assert.assertEquals;
@@ -10,6 +10,9 @@ import static org.junit.Assert.assertNull;
 import org.junit.Test;
 
 import java.awt.Color;
+
+import com.gorman.voxel_engine.world.primitives.Vector;
+import com.gorman.voxel_engine.world.primitives.Voxel;
 
 public class TestChunk
 {
